Finally make the tests pass
Since the new version now uses the official repos from puppetlabs for
apache, the tests needed some additions. It is a good thing that the
actual tests did not need that much changing. That means we are mostly
backwards compatible to the old versions. redmine::params is no longer
needed and the default download method changed from wget to svn. With
rubyforge being down, there is no sensible default download url.