Change the default provider to git
This way the user can just change the version number. Because of the
nature of svn, it is impossible to check out a branch or tag using the
revision directive. We could append the version string to the url, but
that would introduce a consistency just for the svn provider. Instead,
the user will have to pass the full path to the tag or branch to
download. Version should be set to HEAD.