Do bundle update after repository change
When the repository changes, the user has most likely upgraded their
redmine version. As bundle install is only intended to run once on the
first install, we have to use bundle update to install potential new
dependencies and then migrate the database. This bundle update is also
triggered on the first install along with a second db:migrate, but there
is no elegant solution around that and there are no negative side
effects that come from updating or running db:migrate twice.