Actually require all packages before installing
'bundle install' did not require all packages that were necessary to
install the gems. This would result in 'bundle install' failing when
puppet decides to run it before certain packages. Some packages in the
list are not actually required to install all the gems. But keeping a
separate list requires more development effort as the two list have to
be updated constantly and can get out of sync easily.