Fix dependency order when removing plugins
When trying to install a new plugin while also removing one, puppet may
choose the following execution order:
1. Download new plugin
2. Run rake task to remove old plugin
3. ???
4. Fail
Rake fill fail with a Bundler::GemNotFound error because the gems of the
new plugin are not yet installed. WHAT?!
To fix this, the removal must take place AFTER bundle_update. But this
would result in a dependency cycle.
bundle_update => remove plugin => remove directory => bundle_update
The solution is to notify nothing in case of ensure => absent.